p4sync.sh
This simple bash script uses rsync and system built in mail server to syncronize from production server to a backup device or anything that can be mounted on the server. It then emails a summary of what was syncronized.

Optional commented out sections can also prune files, I have since switched to a better script for that.