Create p4log.sh
This script simply truncates the p4d command log on the server and saves it as a new file with that days date, ran nightly with cron this makes it much easier if your system ever has an issue - especially if you need to send your log files to support! I keep 30 days of log files around on the primary server.